requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration
PyYAML>=6.0
pydantic>=2.0.0

# API serving
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0
psutil>=5.9.0
